+++ fjs/media/nix/module.f.ts
@@ -7,15 +7,16 @@
  *
  * @module
  */
-import { type List as ChunkList } from '../../types/list/module.f.ts'
+import { toArray, type List as ChunkList } from '../../types/list/module.f.ts'
 import { concat } from '../../types/string/module.f.ts'
 import { includes } from '../../types/array/module.f.ts'
-import { contains } from '../../types/range/module.f.ts'
 import {
     digitRange,
     latinCapitalLetterRange,
     latinSmallLetterRange,
+    range,
 } from '../../text/ascii/module.f.ts'
+import { fromRange, get, merge } from '../../types/range_set/module.f.ts'
 
 type Identifier = string
 
@@ -69,25 +70,31 @@ const reservedWords = [
 
 const isReservedWord = includes(reservedWords)
 
-const isUpperLetter = contains(...latinCapitalLetterRange)
+const letters = merge
+    (fromRange(latinCapitalLetterRange))
+    (fromRange(latinSmallLetterRange))
 
-const isLowerLetter = contains(...latinSmallLetterRange)
+const identifierInitial = toArray(merge
+    (letters)
+    (fromRange(range('_'))))
 
-const isAsciiLetter = (character: string): boolean => {
-    const c = character.charCodeAt(0)
-    return isUpperLetter(c) || isLowerLetter(c)
-}
-
-const isDigitN = contains(...digitRange)
-
-const isDigit = (character: string): boolean =>
-    isDigitN(character.charCodeAt(0))
+const getIdentifierInitial = get(identifierInitial)
 
 const isIdentifierInitial = (character: string): boolean =>
-    isAsciiLetter(character) || character === '_'
+    getIdentifierInitial(character.charCodeAt(0))
+
+const identifierTrailing = toArray(merge
+    (merge
+        (identifierInitial)
+        (fromRange(digitRange)))
+    (merge
+        (fromRange(range("'")))
+        (fromRange(range('-')))))
+
+const getIdentifierTrailing = get(identifierTrailing)
 
 const isIdentifierTrailing = (character: string): boolean =>
-    isIdentifierInitial(character) || isDigit(character) || character === "'" || character === '-'
+    getIdentifierTrailing(character.charCodeAt(0))
 
 const isIdentifier = (value: string): boolean => {
     const [initial, ...trailing] = value